requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
pymongo
pytube
supabase
youtube-transcript-api
